albedo/relightingDataGen-parallel/requirements.txt
# Core Deep Learning
# Note: Install PyTorch separately with CUDA support:
# conda install pytorch torchvision pytorch-cuda=12.1 -c pytorch -c nvidia
torch>=2.1.0
torchvision>=0.16.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Data handling
pandas>=2.0.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0

# MiDaS Dependencies (for soft shading & specular)
# Required for normal estimation in Stage 3
timm>=0.9.0

# SAM2/SAM3 Dependencies
# Install separately - choose one:
# SAM2 (stable):     pip install git+https://github.com/facebookresearch/sam2.git
# SAM3 (optional):   pip install git+https://github.com/facebookresearch/sam3.git
# Pipeline auto-falls back to SAM2 if SAM3 unavailable

# Note: MiDaS models download automatically via torch.hub on first run
